ENH: added an alternative way to fit stretched exponential
Instead of estimating DDC directly, DDC^alpha is estimated, and then factored out
for the final output.

Note that there is no evidence that this modification will lead to any improvement
of the quality of fit. However, we need to evaluate this approach before discarding
it. In the future, most likely, only one approach to fitting stretched exponential
will be provided in the module.